Fix TestPack and TestUnpack
These tests were calling both Pack() and Unpack() so it was difficult to
determine what was actually happening or if the test was failing because
of Pack() or because of Unpack(). Create a pool of test data that both
tests can utlize. Maybe we need a TestQuine() to make sure that Pack()
and Unpack() are perfect inverses of each other,