Allow users to specify type sizes
Python can use sizeof() to determine native type sizes on the host
system, but Go has no equivalent (atleast without libffi nonsense), so
we allow the user to directly modify the type sizes. This makes the
assumption that the user will know what sizes they want to use. The
standard type sizes defined by the Python documentation are left as
default.

This broke string handling for some reason (although this isn't that
surprising as the string handling was never that robust in the first
place).